Downgrade bluebird to 2.2.2
In production, we saw bursts of memory leaks when encountering cassandra query
errors. One of the largest changes in that deploy was the upgrade of the
bluebird dependency from 2.2.2 to 2.9.30, a version that seems to have a lot
of issues.

This patch rolls back the bluebird upgrade only, so that we can test it in
staging and possibly production.

Change-Id: Ifb82a371eb46e4350d9f835971a57954c9901a26