Config: Consolidate the BetaCluster and production configs into one
We kept the configs separate in ops/puppet, but not a lot of things vary
between all our environments: the UA, the storage groups and domains are
the only differences, so move their definitions into the respective
vars.yaml file.

The domains follow this structure:

  section:
    domain: project

where section is the name of the domain block (like 'Wikipedia' or
'Specials'), domain is the domain name itself and project is the name of
project template to use (which must be defined in config.yaml.j2).

Change-Id: Ia3a989c0994383aab0d502a79f0a517622ba63d2