Skip creating resources on start-up
We have not been using RESTBase to create the needed keyspaces in
Cassandra for a while, so skip this process altogether. Since now it
takes only up to 5 seconds for each worker to bind to the port, reduce
the delay to 12 seconds to allow for at least 9 workers to finish their
start-up sequences before repooling the node.

Change-Id: I7b827815bad3d72640863ab118b680c3baa2045b